fix the long-click behaviour with respect to document and element scrollbars
The underlying use case is to avoid triggering the add-on when the user drags the scrollbar's handle to scroll (which involves long clicking on the scroll handle).

Until now, the strategy has been to determine whether the click occurred on the scrollbar area or on the content area. For the former, the click is ignored. However, the calculation to determine the scrollbar area hasn't been reliable. So even a click in the content area is considered as a click on the scrollbar area on sites like YouTube, failing to trigger long-click even for valid content area.

Now the strategy is to wait for the scroll event to fire from either document or the element on which the click occurred. If the event fires before long-click timeout, then the pinning is not triggered.

As a pleasant side-effect of this fix, the long-click works on sites like YouTube.